+++ package.json
@@ -15,9 +15,6 @@
     "lint": "next lint",
     "start": "next start"
   },
-  "prisma": {
-    "seed": "tsx prisma/seed.ts"
-  },
   "dependencies": {
     "@auth/prisma-adapter": "^1.6.0",
     "@faker-js/faker": "^9.3.0",
