+++ src/app/api/auth/[...nextauth]/options.ts
@@ -1,20 +1,20 @@
-import type { NextAuthOptions } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-
-export const options: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-  ],
-  pages: {
-    signIn: "/login",
-  },
-  callbacks: {
-    async redirect({ baseUrl }) {
-      // Redirect users to the /profile route after login
-      return `${baseUrl}/`;
-    },
-  },
-};
+// import type { NextAuthOptions } from "next-auth";
+// import GoogleProvider from "next-auth/providers/google";
+//
+// export const options: NextAuthOptions = {
+//   providers: [
+//     GoogleProvider({
+//       clientId: process.env.GOOGLE_CLIENT_ID!,
+//       clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+//     }),
+//   ],
+//   pages: {
+//     signIn: "/login",
+//   },
+//   callbacks: {
+//     async redirect({ baseUrl }) {
+//       // Redirect users to the /profile route after login
+//       return `${baseUrl}/`;
+//     },
+//   },
+// };
