+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,5 @@
 import NextAuth from "next-auth";
-import { options } from "./options";
+import { authOptions } from "~/server/auth";
 
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const handler = NextAuth(options);
+const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
